pyproject.toml
[project]
name = "ruty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    
    # LangGraph stack
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    
    # Existing
    "openai>=1.0.0",
    "supermemory>=3.4.0",
    
    # System tray
    "pillow>=12.1.0",
    "pystray>=0.19.5",
]
